Linked footer Send button to Contact Us form. Moved form to the right.
When user was typing in the message text field, the field's scrollbar was covered by the page's scrollbar. Moving the form a bit to the left by increasing right margin fixed the problem.

The button wasn't linked to any form, so I linked it through a created form ID.

Todo: Add form functionality in index.js