Improved register form code and added CSS.
Removed the fieldset and labels, as there was no need

Added CSS drastically improving the form's look, changed the submit input to button and linked it to the form.

Removed extra inputs: gender, nationality, profile picture. They are to be moved to the account update form found in My Account page.

Todo: Add a function which checks if the two password fields match.