Users can now upload avatars, displayed in their profile page.
Users can upload a file through the avatar form on their profile page, which will then be saved in .png format with their username as the filename.

The profile page checks for any images with the name of the currently logged in user, and if found displays their avatar. If not, it displays a default user avatar.