Added delete user function on profile form.
Users can now delete their accounts using the form under Advanced found in the profile edit section of My Account page

Once the checks to verify the user are completed as well as the confirmation checkbox, the user can press the button to delete their account, on which a last confirmation check is given. If the user confirms yet again, their ID is fetched and used as argument in the deleteRecords function, which will check the table and ID after which it will delete the record. The user's avatar file is deleted after.