Added add item form CSS and some HTML.
The description textarea uses jQuery functions to expand automatically while keeping count of the characters in the lower right part. A div is made to look like an input while the textarea has no borders in order to blend in.

There was a bug where the page scrollbar would jump when the textarea reached a certain height, this was fixed by using a different function.