Large update: finished Add Item form structure, design and animations.
Users can upload a maximum of 5 images. If they exceed that number they will get an alert and uploaded images will be reset. When users upload images, thumbnails for them will appear in the box, with zoom-in animations on mouse hover.

Select fields fetch their data from the JSON files, and in order to keep consistency, a js function adjusts the colour of options. In multiple option selects, more options can be selected without the need of Ctrl/Command button hold. There is also a button to clear all selected fields.

Textareas adjust height automatically as user types new lines. A character counter in the bottom right also shows the current characters and maximum field characters in real-time.

After a bit more adjusting of the form, I will work on the server-side implementation.