Improved add item form, modified upper header wrapper styles.
Changed the opacity of file input button to 0 as it wasn't fitting with the form style. Added a text to inform user of how to upload a file (even though it's pretty obvious).

Upper header left and right wrapper are now floated instead of position absolute, so that they don't overlap with other elements when page is zoomed in (with float, the upper-header min-width will take care of that).