Details page image slideshow and more.
Images are now scaled to 16:9, while thumbnails of 1:1 are created in thumbs directory in the listing's image folder.

This is to ensure they are displayed well on the details page gallery slideshow.

Slideshow: Users can view previous and next image by clicking on the buttons on the sides. Under the slideshow, dots are also showing which picture is currently being viewed.

Todo: Replace dots with small thumbnails of each picture, and don't exceed amount of uploaded photos.