Completed: Details Page
Modified CSS so that rows are parallel, added dropdown list for when exchange categories exceed 3, added desired items description.

Added function for Watchlist button, although it is not updating the database yet.

The page view/template is complete, all there is to do now is add buttons functionality and make small adjustments when the need arises