Updated Make Offer page HTML&CSS.
Data of the seller, user, item AND viewing user's listings is sent to the page. Details of the current item are shown, as well as the seller's name and avatar. The viewing user can make an offer by sending a private message and also has the option of attaching one of his/her listings through a Select form (TODO: DESIGN).

After the select design is done, I will work on the back-end.