Offers are now sent to seller's email address.
The offer contains the user who made the offer, their item (if they have selected one) put for exchange, and the short message.

Todo: Email design and links to items, as well as the pictures. Basically make it look like a real email you'd normally get from eBay or something about an offer.