Added listing delete button when user is viewing own listing.
Added checks to see if viewing user ID is same as seller id (to know if user is viewing own listing). If this is true, use handlebars to display a one-button form to delete the listing. Function removeItem in accounts.js will then do another check to see if user requesting the delete is the seller, and if that is the case, call the deleteRecord function to erase the listing from the database.