Parameterized all database methods in accounts.js
This is to protect against SQL Injections.

One exception made would be the fetchUserWatchListings function, which looks for watchlist items using 'IN'. It is harder to pass data to this SQL operator, and the function is used only by other functions which have been protected, and not directly by the user, which is why it was omitted.

Further testing needs to be done to ensure this big change didn't break any working code.