Fixed multiple bugs created by SQL query parameterization
User sign up date wouldn't save in the correct format in the database, making it a string solved the issue.

When adding a new item, selecting multiple desired item categories would break the listing, as the categories would be saved as null since the body considered it an array. Joining all array elements into a string when giving it as a parameter to the database query seems to have fixed the issue.