Fixed more bugs, user accounts delete now removes listings too.
There were more bugs with functions such as fetch and add items, specifically ExchangeCategories not being properly saved in the database when multiple were selected, as it made the variable be an Object instead of a string. Added code to convert to string when that happens so SQL can accept it.

When the user deletes his/her account, their listings will also be deleted from the database along with their files (images and thumbnails).