Fixed listing bugs, check description
When there would be no other listings saved in the Database in items, the sequence would be null, and so the item ID wouldn't get passed right for the images saving function. This was fixed by checking if the item id is undefined, and if so, set it as 1 in the add-item script.

Another issue would be displaying listings on the main page when there would be only one or none saved. This was due to the data being passed to Handlebars as an object instead of an Array. Added conditional checks to see if it is a single object, in which case it converts it to an object inside an array; and another check for if no items are passed, in which case it sets items to undefined (so Handlebars can recognize it is empty).